restricted pages

+++ base/views.py
@@ -1,12 +1,17 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.http import HttpResponse
 from .models import Room, Topic, User
 from .forms import RoomForm
 
 def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -53,6 +58,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
    
@@ -65,9 +71,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user == room.host:
+        return HttpResponse('You are not allowed here!!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -78,8 +88,12 @@ def updateRoom(request, pk):
     context = {'form': form, 'room': room}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user == room.host:
+        return HttpResponse('You are not allowed here!!')
     
     if request.method == 'POST':
         room.delete()
